requirements.txt
numpy==2.4.2
pandas>=2.0,<3.0
kagglehub==1.0.0
joblib==1.5.2

scikit_learn==1.8.0
lightgbm==4.6.0
xgboost==3.2.0

mlflow==3.10.0
mlflow_skinny==3.10.0
mlflow_tracing==3.10.0
dagshub==0.6.5

fastapi==0.132.0
pydantic==2.12.5
python-dotenv==1.2.1
uvicorn==0.40.0